unnecessary casting

+++ RunCat/Program.cs
@@ -329,7 +329,7 @@ namespace RunCat
         {
             if (!speed.Equals("default"))
             {            
-                float manualInterval = (float)Math.Max(minCPU, interval);
+                float manualInterval = Math.Max(minCPU, interval);
                 animateTimer.Stop();
                 animateTimer.Interval = (int)manualInterval;
                 animateTimer.Start();
